dm: call blk_revalidate_disk_zones to clean up zoned settings
If a dm device switches from being zoned to non-zoned during a table
swap, it should clean up the zoned settings from its gendisk.
blk_revalidate_disk_zones() will do this, but it can only be called if
it is possible to freeze the queue, otherwise it will hang. This
means that it can only be called when there are no write plugged zone
bios outstanding.