dm: fix dm_blk_report_zones
If dm_get_live_table() returned NULL, dm_put_live_table() was never
called.  Also, if md->zone_revalidate_map is set, there is no guarantee
that dm_blk_report_zones() is being called from the process that set it
in __bind(). In that case, we need to guard against zone_revalidate_map,
the new map being swapped in, getting free if __bind() fails while we
are in this call. dm_resume() will wait for the SRCU to complete before
freeing the new map on failure.