dm-mpath: Don't grab work_mutex while probing paths
Grabbing the work_mutex keeps probe_active_paths() from running at the
same time as multipath_message(). The only messages that could interfere
with probing the paths are "disable_group", "enable_group", and
"switch_group". These messages could force multipath to pick a new
pathgroup while probe_active_paths() was probing the current pathgroup.
If the multipath device has a hardware handler, and it switches active
pathgroups while there is outstanding IO to a path device, it's possible
that IO to the path will fail, even if the path would be usable if it
was in the active pathgroup. To avoid this, do not clear the current
pathgroup for the *_group messages while probe_active_paths() is
running. Instead set a flag, and probe_active_paths() will clear it
when it finishes probing the paths.

This works fine for disable_group and enable_group, since setting
pg->bypassed has no effect until the current pathgroup is cleared or has
no usable paths. But switch_group sets next_pg, which gets checked
before current_pg. However, there's no need for it to be checked in that
order. Before this patch, next_pg is only ever set when current_pg is
NULL. So this patch makes multipath check current_pg for a usable path
before checking next_pg. If we are waiting for probe_active_paths() to
finish before clearing current_pg, next_pg will be ignored as long as
current_pg has usable paths, like intended.

While this patch will not change the behavior of existing multipath
users (which don't use the DM_MPATH_PROBE_PATHS ioctl), when that ioctl
is used, the behavior of the "disable_group", "enable_group", and
"switch_group" messages can change subtly. When these messages return,
the next IO to the multipath device will no longer be guaranteed to
choose a new pathgroup. Instead, choosing a new pathgroup could be
delayed by an in-progress DM_MPATH_PROBE_PATHS ioctl. The userspace
multipath tools make no assumptions about what will happen to IOs after
sending these messages, so this change will not effect already released
versions of them, even if the DM_MPATH_PROBE_PATHS ioctl is run
alongside them.